Social Media/requirements.txt
customtkinter>=5.2.0
aiohttp>=3.8.0
asyncio
typing-extensions>=4.0.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
librosa>=0.10.0
numpy>=1.24.0
psutil>=5.9.0
multiprocessing-logging>=0.3.0
instagrapi>=2.0.0
tweepy>=4.14.0
linkedin-api>=2.0.0
facebook-sdk>=3.1.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pydub>=0.25.1
librosa>=0.10.0
mutagen>=1.47.0
schedule>=1.2.0